backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP and API
aiohttp==3.9.1
requests==2.31.0

# File Handling
aiofiles==23.2.1

# Image Processing and OCR
pillow==10.1.0
PyMuPDF==1.23.8

# AI and Machine Learning
pyautogen==0.2.3
openai==1.3.7
groq==0.4.1

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
